# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Unifiles API
  description: Canonical contract for the self-hosted Unifiles document processing API. Server bindings
    and official SDK cores are generated from this document.
  version: 1.0.0
paths:
...
  /v1/files:
    post:
      tags:
      - Files
      summary: Upload File
      operationId: uploadFile
      security:
      - BearerAuth: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          title: Idempotency-Key
          nullable: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadFileRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags:
      - Files
      summary: List Files
      operationId: listFiles
      security:
      - BearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: tags
        in: query
        required: false
        schema:
          type: string
          title: Tags
          nullable: true
      - name: content_type
        in: query
        required: false
        schema:
          type: string
          title: Content Type
          nullable: true
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          default: created_at
          title: Sort By
      - name: order
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Order
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /v1/usage/stats:
    get:
      tags:
      - Usage
      summary: Usage Stats
      operationId: getUsageStats
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageStatsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - BearerAuth: []
  /v1/usage/limits:
    get:
      tags:
      - Usage
      summary: Usage Limits
      operationId: getUsageLimits
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageLimitsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - BearerAuth: []
components:
  schemas:
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
        request_id:
          type: string
          title: Request Id
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - request_id
      title: ErrorDetail
    ErrorEnvelope:
      properties:
        success:
          type: boolean
          title: Success
          default: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorEnvelope
...
    FileResource:
      properties:
        id:
          type: string
          title: Id
        filename:
          type: string
          title: Filename
        content_type:
          type: string
          title: Content Type
        size:
          type: integer
          title: Size
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        tags:
          items:
            type: string
          type: array
          title: Tags
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
          nullable: true
      additionalProperties: false
      type: object
      required:
      - id
      - filename
      - content_type
      - size
      - created_at
      title: FileResource
...
    UploadFileRequest:
      properties:
        file:
          type: string
          format: binary
          title: File
        metadata:
          type: string
          title: Metadata
          default: '{}'
        tags:
          type: string
          title: Tags
          default: '[]'
      type: object
      required:
      - file
      title: UploadFileRequest
...
    FileList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FileResource'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      additionalProperties: false
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: FileList
...
    FileResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/FileResource'
      additionalProperties: false
      type: object
      required:
      - data
      title: FileResponse
...
    FileListResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/FileList'
      additionalProperties: false
      type: object
      required:
      - data
      title: FileListResponse
...
    StorageUsage:
      type: object
      additionalProperties: false
      properties:
        used_bytes:
          type: integer
        limit_bytes:
          type: integer
        used_percentage:
          type: number
      required:
      - used_bytes
      - limit_bytes
      - used_percentage
      title: StorageUsage
    ExtractionUsage:
      type: object
      additionalProperties: false
      properties:
        pages_used:
          type: integer
        pages_limit:
          type: integer
        reset_at:
          type: string
          format: date-time
          nullable: true
      required:
      - pages_used
      - pages_limit
      title: ExtractionUsage
    KnowledgeBaseUsage:
      type: object
      additionalProperties: false
      properties:
        used:
          type: integer
        limit:
          type: integer
      required:
      - used
      - limit
      title: KnowledgeBaseUsage
    UsageStats:
      type: object
      additionalProperties: false
      properties:
        storage:
          $ref: '#/components/schemas/StorageUsage'
        extraction:
          $ref: '#/components/schemas/ExtractionUsage'
        knowledge_bases:
          $ref: '#/components/schemas/KnowledgeBaseUsage'
      required:
      - storage
      - extraction
      - knowledge_bases
      title: UsageStats
    ApiCallLimits:
      type: object
      additionalProperties: false
      properties:
        limit:
          type: integer
          nullable: true
        window:
          type: string
      required:
      - limit
      - window
      title: ApiCallLimits
    StorageLimits:
      type: object
      additionalProperties: false
      properties:
        limit_bytes:
          type: integer
      required:
      - limit_bytes
      title: StorageLimits
    FileLimits:
      type: object
      additionalProperties: false
      properties:
        max_file_size_bytes:
          type: integer
      required:
      - max_file_size_bytes
      title: FileLimits
    UsageLimits:
      type: object
      additionalProperties: false
      properties:
        api_calls:
          $ref: '#/components/schemas/ApiCallLimits'
        storage:
          $ref: '#/components/schemas/StorageLimits'
        files:
          $ref: '#/components/schemas/FileLimits'
      required:
      - api_calls
      - storage
      - files
      title: UsageLimits
...
    UsageStatsResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/UsageStats'
      required:
      - data
      title: UsageStatsResponse
    UsageLimitsResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/UsageLimits'
      required:
      - data
      title: UsageLimitsResponse
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
servers:
- url: https://api.unifiles.dev
  description: Hosted API
- url: http://localhost:8088
  description: Local self-hosted API
...