# apps/server/generated/src/unifiles_server_protocol/models/usage_stats.py
# coding: utf-8

"""
    Unifiles API

    Canonical contract for the self-hosted Unifiles document processing API. Server bindings and official SDK cores are generated from this document.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict
from typing import Any, ClassVar, Dict
from unifiles_server_protocol.models.extraction_usage import ExtractionUsage
from unifiles_server_protocol.models.knowledge_base_usage import KnowledgeBaseUsage
from unifiles_server_protocol.models.storage_usage import StorageUsage
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class UsageStats(BaseModel):
    """
    UsageStats
    """ # noqa: E501
    storage: StorageUsage
    extraction: ExtractionUsage
    knowledge_bases: KnowledgeBaseUsage
    __properties: ClassVar[List[str]] = ["storage", "extraction", "knowledge_bases"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of UsageStats from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of storage
        if self.storage:
            _dict['storage'] = self.storage.to_dict()
        # override the default output from pydantic by calling `to_dict()` of extraction
        if self.extraction:
            _dict['extraction'] = self.extraction.to_dict()
        # override the default output from pydantic by calling `to_dict()` of knowledge_bases
        if self.knowledge_bases:
            _dict['knowledge_bases'] = self.knowledge_bases.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of UsageStats from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "storage": StorageUsage.from_dict(obj.get("storage")) if obj.get("storage") is not None else None,
            "extraction": ExtractionUsage.from_dict(obj.get("extraction")) if obj.get("extraction") is not None else None,
            "knowledge_bases": KnowledgeBaseUsage.from_dict(obj.get("knowledge_bases")) if obj.get("knowledge_bases") is not None else None
        })
        return _obj